requirements.txt
google-generativeai
arxiv
PyPDF2